requirements.txt
flask==3.1.0
requests==2.32.3
python-dotenv==1.1.0
beautifulsoup4==4.13.3
gunicorn==23.0.0
